requirements.txt
Django==3.1.4
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
django-rest-swagger==2.2.0
django-cors-headers==3.2.0
drf-yasg==1.20.0
djoser==2.1.0